requirements.txt
blinker
certifi  
charset-normalizer
click
colorama
crypto
distlib
et-xmlfile
exceptiongroup
filelock    
Flask
Flask-SQLAlchemy
Flask-WTF
greenlet
gunicorn
idna
importlib_metadata
iniconfig
itsdangerous
Jinja2
MarkupSafe
mysql-connector
Naked
openpyxl
packaging
pip
python-dotenv
platformdirs
pluggy
pycryptodome
pytest
PyYAML
requests
setuptools
shellescape
SQLAlchemy
tomli
typing_extensions
urllib3
virtualenv
Werkzeug
WTForm
WTForms
zipp                  
python-dotenv
pymysql